refactor: split up actions-bar and more-menu
