+++ src/components/DashboardsBarNew/InformationBlock/ActionsBar.js
@@ -1,180 +1,38 @@
 import { OfflineTooltip } from '@dhis2/analytics'
 import { useDhis2ConnectionStatus } from '@dhis2/app-runtime'
 import i18n from '@dhis2/d2-i18n'
-import {
-    Button,
-    FlyoutMenu,
-    colors,
-    IconMore16,
-    SharingDialog,
-} from '@dhis2/ui'
+import { Button, colors, IconMore16, SharingDialog } from '@dhis2/ui'
 import PropTypes from 'prop-types'
 import React, { useCallback, useState } from 'react'
 import { connect } from 'react-redux'
-import { useHistory, Redirect } from 'react-router-dom'
-import { acClearItemFilters } from '../../../actions/itemFilters.js'
-import { acSetShowDescription } from '../../../actions/showDescription.js'
-import { apiPostShowDescription } from '../../../api/description.js'
-import { useCacheableSection } from '../../../modules/useCacheableSection.js'
+import { useHistory } from 'react-router-dom'
 import { orObject } from '../../../modules/util.js'
-import { ROUTE_START_PATH } from '../../../pages/start/index.js'
-import { sGetNamedItemFilters } from '../../../reducers/itemFilters.js'
 import { sGetSelected } from '../../../reducers/selected.js'
-import { sGetShowDescription } from '../../../reducers/showDescription.js'
-import ConfirmActionDialog from '../../ConfirmActionDialog.js'
 import DropdownButton from '../../DropdownButton/DropdownButton.js'
-import MenuItem from '../../MenuItemWithTooltip.js'
 import FilterSelector from './FilterSelector.js'
+import MoreOptionsMenu from './MoreOptionsMenu.js'
 import classes from './styles/ActionsBar.module.css'
 
-const ViewActions = ({
-    id,
+const ActionsBar = ({
     access,
-    showDescription,
+    allowedFilters,
+    id,
+    restrictFilters,
+    showAlert,
     starred,
     toggleDashboardStarred,
-    showAlert,
-    updateShowDescription,
-    removeAllFilters,
-    restrictFilters,
-    allowedFilters,
-    filtersLength,
 }) => {
     const history = useHistory()
     const [moreOptionsIsOpen, setMoreOptionsIsOpen] = useState(false)
     const [sharingDialogIsOpen, setSharingDialogIsOpen] = useState(false)
-    const [confirmCacheDialogIsOpen, setConfirmCacheDialogIsOpen] =
-        useState(false)
-    const [redirectUrl, setRedirectUrl] = useState(null)
     const { isDisconnected: offline } = useDhis2ConnectionStatus()
-    const { lastUpdated, isCached, startRecording, remove } =
-        useCacheableSection(id)
-
-    const onRecordError = useCallback(() => {
-        showAlert({
-            msg: i18n.t(
-                "The dashboard couldn't be made available offline. Try again."
-            ),
-            isCritical: true,
-        })
-    }, [showAlert])
-
-    const onCacheDashboardConfirmed = useCallback(() => {
-        setConfirmCacheDialogIsOpen(false)
-        removeAllFilters()
-        startRecording({
-            onError: onRecordError,
-        })
-    }, [onRecordError, removeAllFilters, startRecording])
-
-    const onRemoveFromOffline = useCallback(() => {
+    const onToggleSharingDialog = useCallback(() => {
+        setSharingDialogIsOpen((currentOpenState) => !currentOpenState)
+    }, [])
+    const closeMoreOptionsMenu = useCallback(() => {
         setMoreOptionsIsOpen(false)
-        lastUpdated && remove()
-    }, [lastUpdated, remove])
-
-    const onAddToOffline = useCallback(() => {
-        setMoreOptionsIsOpen(false)
-        return filtersLength
-            ? setConfirmCacheDialogIsOpen(true)
-            : startRecording({
-                  onError: onRecordError,
-              })
-    }, [filtersLength, onRecordError, startRecording])
-
-    const onToggleShowDescription = useCallback(() => {
-        updateShowDescription(!showDescription)
-        setMoreOptionsIsOpen(false)
-        !offline && apiPostShowDescription(!showDescription)
-    }, [offline, showDescription, updateShowDescription])
-
-    const onToggleSharingDialog = useCallback(
-        () => setSharingDialogIsOpen(!sharingDialogIsOpen),
-        [sharingDialogIsOpen]
-    )
-
+    }, [])
     const userAccess = orObject(access)
-
-    const getMoreMenu = () => (
-        <FlyoutMenu>
-            {lastUpdated ? (
-                <MenuItem
-                    dense
-                    disabledWhenOffline={false}
-                    label={i18n.t('Remove from offline storage')}
-                    onClick={onRemoveFromOffline}
-                />
-            ) : (
-                <MenuItem
-                    dense
-                    disabled={offline}
-                    label={i18n.t('Make available offline')}
-                    onClick={onAddToOffline}
-                />
-            )}
-            {lastUpdated && (
-                <MenuItem
-                    dense
-                    label={i18n.t('Sync offline data now')}
-                    disabled={offline}
-                    onClick={onAddToOffline}
-                />
-            )}
-            <MenuItem
-                dense
-                disabled={offline}
-                label={
-                    starred
-                        ? i18n.t('Unstar dashboard')
-                        : i18n.t('Star dashboard')
-                }
-                onClick={toggleDashboardStarred}
-            />
-            <MenuItem
-                dense
-                disabledWhenOffline={false}
-                label={
-                    showDescription
-                        ? i18n.t('Hide description')
-                        : i18n.t('Show description')
-                }
-                onClick={onToggleShowDescription}
-            />
-            <MenuItem
-                dense
-                disabled={offline && !isCached}
-                disabledWhenOffline={false}
-                label={i18n.t('Print')}
-                dataTest="print-menu-item"
-            >
-                <MenuItem
-                    dense
-                    disabled={offline && !isCached}
-                    disabledWhenOffline={false}
-                    label={i18n.t('Dashboard layout')}
-                    onClick={() => setRedirectUrl(`${id}/printlayout`)}
-                    dataTest="print-layout-menu-item"
-                />
-                <MenuItem
-                    dense
-                    disabled={offline && !isCached}
-                    disabledWhenOffline={false}
-                    label={i18n.t('One item per page')}
-                    onClick={() => setRedirectUrl(`${id}/printoipp`)}
-                    dataTest="print-oipp-menu-item"
-                />
-            </MenuItem>
-            <MenuItem
-                dense
-                disabledWhenOffline={false}
-                label={i18n.t('Close dashboard')}
-                onClick={() => history.push(ROUTE_START_PATH)}
-            />
-        </FlyoutMenu>
-    )
-
-    if (redirectUrl) {
-        return <Redirect to={redirectUrl} />
-    }
 
     return (
         <>
@@ -186,7 +44,7 @@ const ViewActions = ({
                             small
                             disabled={offline}
                             className={classes.editButton}
-                            onClick={() => setRedirectUrl(`${id}/edit`)}
+                            onClick={() => history.push(`${id}/edit`)}
                         >
                             {i18n.t('Edit')}
                         </Button>
@@ -217,7 +75,14 @@ const ViewActions = ({
                     disabledWhenOffline={false}
                     onClick={() => setMoreOptionsIsOpen(!moreOptionsIsOpen)}
                     icon={<IconMore16 color={colors.grey700} />}
-                    component={getMoreMenu()}
+                    component={
+                        <MoreOptionsMenu
+                            close={closeMoreOptionsMenu}
+                            showAlert={showAlert}
+                            toggleDashboardStarred={toggleDashboardStarred}
+                            starred={starred}
+                        />
+                    }
                 >
                     <wbr />
                 </DropdownButton>
@@ -229,46 +94,29 @@ const ViewActions = ({
                     onClose={onToggleSharingDialog}
                 />
             )}
-            <ConfirmActionDialog
-                title={i18n.t('Clear dashboard filters?')}
-                message={i18n.t(
-                    "A dashboard's filters can’t be saved offline. Do you want to remove the filters and make this dashboard available offline?"
-                )}
-                cancelLabel={i18n.t('No, cancel')}
-                confirmLabel={i18n.t('Yes, clear filters and sync')}
-                onConfirm={onCacheDashboardConfirmed}
-                onCancel={() => setConfirmCacheDialogIsOpen(false)}
-                open={confirmCacheDialogIsOpen}
-            />
         </>
     )
 }
 
-ViewActions.propTypes = {
+ActionsBar.propTypes = {
     access: PropTypes.object,
     allowedFilters: PropTypes.array,
-    filtersLength: PropTypes.number,
     id: PropTypes.string,
-    removeAllFilters: PropTypes.func,
     restrictFilters: PropTypes.bool,
     showAlert: PropTypes.func,
-    showDescription: PropTypes.bool,
     starred: PropTypes.bool,
     toggleDashboardStarred: PropTypes.func,
-    updateShowDescription: PropTypes.func,
 }
 
 const mapStateToProps = (state) => {
     const dashboard = sGetSelected(state)
 
     return {
-        ...dashboard,
-        filtersLength: sGetNamedItemFilters(state).length,
-        showDescription: sGetShowDescription(state),
+        access: dashboard.access,
+        allowedFilters: dashboard.allowedFilters,
+        id: dashboard.id,
+        restrictFilters: dashboard.restrictFilters,
     }
 }
 
-export default connect(mapStateToProps, {
-    removeAllFilters: acClearItemFilters,
-    updateShowDescription: acSetShowDescription,
-})(ViewActions)
+export default connect(mapStateToProps)(ActionsBar)
